Adding canned messages for Donot Offer Messages
