+++ app/routes/offer/donor_messages.js
@@ -1,8 +1,14 @@
-import MessagesBaseRoute from 'shared-goodcity/routes/messages_base';
+import MessagesBaseRoute from "../message_base";
 
 export default MessagesBaseRoute.extend({
   afterModel() {
     var offerId = this.modelFor("offer").get("id");
-    this.store.query('version', { item_id: offerId, for_offer: true });
+    this.store.query("version", { item_id: offerId, for_offer: true });
+  },
+
+  resetController(controller, isExiting) {
+    if (isExiting) {
+      controller.set("isActive", false);
+    }
   }
 });
