+++ app/styles/templates/components/_canned-messages-overlay.scss
@@ -1,6 +1,7 @@
 .scaled_icon {
   transform: scale(2);
   margin-top: 0.8rem;
+  color: black;
 }
 
 .canned-search-container {
